# api/openapi.yaml
openapi: 3.0.3
info:
  title: Jupiter API
  version: "2026-09-30"
  description: |
    Jupiter's merchant API. Every response describes the version it was rendered in with
    the Jupiter-Version header; this document describes the current version.

    Authenticate with a secret or restricted key as a bearer token. Keys prefixed
    `_test_` act on test-mode data and `_live_` on live-mode data; the two never mix.

    Every POST accepts an Idempotency-Key header. A retry with the same key and the same
    body returns the first response without acting again; the same key with a different
    body is refused.
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/risk/decisions:
    get:
      operationId: listRiskDecisions
      tags: [Risk]
      description: The decision log, newest first. Every decision lists the rules that fired and the features they saw.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/StartingAfter"
        - $ref: "#/components/parameters/EndingBefore"
        - name: action
          in: query
          schema: { type: string, enum: [allow, review, request_3ds, block] }
        - name: payment_intent
          in: query
          schema: { type: string }
      responses:
        "200":
          description: A page of decisions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskDecisionList" }
        default: { $ref: "#/components/responses/Failure" }
  /v1/risk/decisions/{id}:
    get:
      operationId: getRiskDecision
      tags: [Risk]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: The decision.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskDecision" }
        default: { $ref: "#/components/responses/Failure" }
  /v1/risk/rules:
    get:
      operationId: listRiskRules
      tags: [Risk]
      description: The merchant's own rules, and the platform's, which apply to every payment.
      responses:
        "200":
          description: The rules.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskRulePage" }
        default: { $ref: "#/components/responses/Failure" }
    post:
      operationId: createRiskRule
      tags: [Risk]
      description: |
        Adds a rule: when expression holds for a payment, action applies. Expressions use
        the features a decision records, such as `amount >= 500000 && brand == "elo"` or
        `ip_attempts_1h > 3`; docs/api/risk.md lists them.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRiskRuleRequest" }
      responses:
        "200":
          description: The rule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskRule" }
        default: { $ref: "#/components/responses/Failure" }
  /v1/risk/rules/{id}:
    delete:
      operationId: deleteRiskRule
      tags: [Risk]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: The deleted rule.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletedObject" }
        default: { $ref: "#/components/responses/Failure" }
  /v1/risk/list_items:
    get:
      operationId: listRiskListItems
      tags: [Risk]
      responses:
        "200":
          description: The allow and block list entries.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskListItemPage" }
        default: { $ref: "#/components/responses/Failure" }
    post:
      operationId: createRiskListItem
      tags: [Risk]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRiskListItemRequest" }
      responses:
        "200":
          description: The entry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RiskListItem" }
        default: { $ref: "#/components/responses/Failure" }
  /v1/risk/list_items/{id}:
    delete:
      operationId: deleteRiskListItem
      tags: [Risk]
      parameters:
        - $ref: "#/components/parameters/ID"
      responses:
        "200":
          description: The deleted entry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeletedObject" }
        default: { $ref: "#/components/responses/Failure" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
    StartingAfter:
      name: starting_after
      in: query
      description: An object id; the page starts with the next older object.
      schema: { type: string }
    EndingBefore:
      name: ending_before
      in: query
      description: An object id; the page ends with the next newer object.
      schema: { type: string }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, maxLength: 255 }
  responses:
    Failure:
      description: The request failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/Error" }
    Error:
      type: object
      required: [type, code, message, request_id, doc_url]
      properties:
        type:
          type: string
          enum: [api_error, authentication_error, card_error, permission_error, invalid_request_error, idempotency_error, rate_limit_error]
        code: { type: string }
        decline_code: { type: string }
        payment_intent: { $ref: "#/components/schemas/PaymentIntent" }
        message: { type: string }
        param: { type: string }
        request_id: { type: string }
        doc_url: { type: string }
    ListMeta:
      type: object
      required: [object, has_more, url]
      properties:
        object: { type: string, enum: [list] }
        has_more: { type: boolean }
        url: { type: string }
...
    PaymentIntent:
      type: object
      required: [id, object, livemode, amount, currency, capture_method, status, payment_method, description,
                 amount_capturable, amount_received, amount_refunded, latest_attempt, last_payment_error,
                 next_action, cancellation_reason, installments, setup_future_usage, request_three_d_secure,
                 risk_decision, created]
      properties:
        id: { type: string }
        object: { type: string, enum: [payment_intent] }
        livemode: { type: boolean }
        amount: { type: integer, format: int64, description: "In the currency's minor unit, such as centavos." }
        currency: { type: string, description: "Lower-case ISO 4217 code." }
        capture_method: { type: string, enum: [automatic, manual] }
        status:
          type: string
          enum: [requires_payment_method, requires_confirmation, requires_action, processing, requires_capture, succeeded, canceled]
        payment_method: { type: string, nullable: true }
        description: { type: string }
        amount_capturable: { type: integer, format: int64 }
        amount_received: { type: integer, format: int64 }
        amount_refunded: { type: integer, format: int64 }
        latest_attempt: { type: string, nullable: true }
        last_payment_error:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/PaymentError" }]
        next_action:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/NextAction" }]
        cancellation_reason: { type: string, nullable: true }
        installments:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Installments" }]
        setup_future_usage: { type: string, enum: [off_session], nullable: true }
        request_three_d_secure: { type: string, enum: [automatic, any] }
        risk_decision:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/RiskDecisionRef" }]
        created: { type: integer, format: int64 }
    RiskDecisionRef:
      type: object
      required: [id, action]
      properties:
        id: { type: string }
        action: { type: string, enum: [allow, review, request_3ds, block] }
    Installments:
      type: object
      description: |
        Parcelado: the payment is authorized once, for the whole amount, and the cardholder
        pays it in count monthly installments. financed_by merchant is parcelado lojista,
        without interest to the cardholder; issuer is parcelado emissor, with interest.
      required: [count, financed_by]
      additionalProperties: false
      properties:
        count: { type: integer, minimum: 2, maximum: 12 }
        financed_by: { type: string, enum: [merchant, issuer] }
    PaymentError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        decline_code: { type: string }
        message: { type: string }
    RiskDecision:
      type: object
      required: [id, object, livemode, payment_intent, attempt, action, rules, features, created]
      properties:
        id: { type: string }
        object: { type: string, enum: [risk_decision] }
        livemode: { type: boolean }
        payment_intent: { type: string }
        attempt: { type: string }
        action: { type: string, enum: [allow, review, request_3ds, block] }
        rules:
          type: array
          description: The rules and list entries that fired; the decision is the most severe of their actions.
          items: { $ref: "#/components/schemas/FiredRule" }
        features:
          type: object
          description: What the rules saw.
          additionalProperties: true
        created: { type: integer, format: int64 }
    FiredRule:
      type: object
      required: [id, action, description]
      properties:
        id: { type: string }
        action: { type: string, enum: [allow, review, request_3ds, block] }
        description: { type: string }
        expression: { type: string }
    RiskDecisionList:
      allOf:
        - $ref: "#/components/schemas/ListMeta"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/RiskDecision" }
    RiskRule:
      type: object
      required: [id, object, action, expression, description, platform]
      properties:
        id: { type: string }
        object: { type: string, enum: [risk_rule] }
        action: { type: string, enum: [allow, review, request_3ds, block] }
        expression: { type: string }
        description: { type: string }
        platform: { type: boolean, description: "A platform rule, which applies to every merchant and cannot be deleted." }
        created: { type: integer, format: int64, nullable: true }
    RiskRulePage:
      type: object
      required: [object, data]
      properties:
        object: { type: string, enum: [list] }
        data:
          type: array
          items: { $ref: "#/components/schemas/RiskRule" }
    CreateRiskRuleRequest:
      type: object
      required: [action, expression]
      additionalProperties: false
      properties:
        action: { type: string, enum: [allow, review, request_3ds, block] }
        expression: { type: string, maxLength: 1000 }
        description: { type: string, maxLength: 500 }
    RiskListItem:
      type: object
      required: [id, object, list, kind, value, created]
      properties:
        id: { type: string }
        object: { type: string, enum: [risk_list_item] }
        list: { type: string, enum: [allow, block] }
        kind: { type: string, enum: [card_fingerprint, ip, bin] }
        value: { type: string }
        created: { type: integer, format: int64 }
    RiskListItemPage:
      type: object
      required: [object, data]
      properties:
        object: { type: string, enum: [list] }
        data:
          type: array
          items: { $ref: "#/components/schemas/RiskListItem" }
    CreateRiskListItemRequest:
      type: object
      required: [list, kind, value]
      additionalProperties: false
      properties:
        list: { type: string, enum: [allow, block] }
        kind: { type: string, enum: [card_fingerprint, ip, bin] }
        value: { type: string, maxLength: 100 }
    NextAction:
      type: object
      required: [type]
      properties:
        type:
          type: string
          description: |
            redirect_to_url: send the customer to redirect_to_url.url to complete 3-D Secure.
            use_test_authentication: in test mode, complete it with the test helper.
        redirect_to_url:
          type: object
          required: [url]
          properties:
            url: { type: string }
...
    DeletedObject:
      type: object
      required: [id, object, deleted]
      properties:
        id: { type: string }
        object: { type: string }
        deleted: { type: boolean, enum: [true] }
...