# internal/risk/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type RiskAttempt struct {
	AttemptID       string
	MerchantID      string
	Livemode        bool
	CardFingerprint string
	Ip              string
	CreatedAt       pgtype.Timestamptz
	Outcome         string
}

type RiskDecision struct {
	ID         string
	AttemptID  string
	IntentID   string
	MerchantID string
	Livemode   bool
	Action     string
	Rules      []byte
	Features   []byte
	CreatedAt  pgtype.Timestamptz
}

type RiskListItem struct {
	ID         string
	MerchantID string
	Livemode   bool
	List       string
	Kind       string
	Value      string
	CreatedAt  pgtype.Timestamptz
}

type RiskRule struct {
	ID          string
	MerchantID  string
	Livemode    bool
	Action      string
	Expression  string
	Description string
	CreatedAt   pgtype.Timestamptz
	DeletedAt   pgtype.Timestamptz
}

type RiskThrottle struct {
	MerchantID string
	Livemode   bool
	StartedAt  pgtype.Timestamptz
	Until      pgtype.Timestamptz
	Ratio      float64
	Attempts   int32
}